pyproject.toml
[project]
name = "jira-evidence-platform"
version = "0.1.0"
description = "Agent-orchestrated Jira evidence automation platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-slugify>=8.0.4",
  "uvicorn>=0.30.0",
  "playwright>=1.46.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
jira-evidence-cli = "cli.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [
  "apps/api/src",
  "apps/cli/src",
  "packages/agent_runtime/src",
  "packages/browser_automation/src",
  "packages/evidence_service/src",
  "packages/jira_mcp/src",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
  "apps/api/src",
  "apps/cli/src",
  "packages/agent_runtime/src",
  "packages/browser_automation/src",
  "packages/evidence_service/src",
  "packages/jira_mcp/src",
]
